ajout: suppression des avis en admin

+++ App/controllers/DeleteUserController.php
@@ -1,11 +1,17 @@
 <?php
-class DeleteUserController{
+class DeleteUserController
+{
 
-    public function deleteUser($id){
-        require_once __DIR__ . '/../models/Database.php';
-        $db = new Database();
-        if ($db->deleteUser($id)) {
-            header("Location: /detailsUtilisateur");
+    public function deleteUser($id)
+    {
+        if (isset($_SESSION['isAdmin'])) {
+            require_once __DIR__ . '/../models/Database.php';
+            $db = new Database();
+            if ($db->deleteUser($id)) {
+                header("Location: /detailsUtilisateur");
+            }
+        } else {
+            header('Location: /');
         }
     }
 }
